# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SsuBench API
  version: "1.0.0"
  description: REST API specification for the current implemented SsuBench endpoints.
servers:
  - url: http://localhost:8080
tags:
  - name: Auth
  - name: Users
  - name: Tasks
  - name: Bids
  - name: Payments

security:
  - bearerAuth: [ ]

paths:
  /auth/register:
    post:
      security: []
      tags: [Auth]
      summary: Register a new user
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
        - in: query
          name: password
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: role
          required: true
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/login:
    post:
      security: []
      tags: [Auth]
      summary: Login with email and password
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
        - in: query
          name: password
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /users:
    get:
      tags: [Users]
      summary: List users with pagination
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Users]
      summary: Create user
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
        - in: query
          name: password
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: role
          required: true
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
        - in: query
          name: password
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: role
          required: true
          schema:
            $ref: '#/components/schemas/Role'
        - in: query
          name: balance
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/{id}/balance:
    post:
      tags: [Users]
      summary: Update user balance
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - in: query
          name: balance
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/{id}/block:
    post:
      tags: [Users]
      summary: Block user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Blocked
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/{id}/unblock:
    post:
      tags: [Users]
      summary: Unblock user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Unblocked
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks with pagination
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Tasks]
      summary: Create task
      parameters:
        - in: query
          name: title
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: description
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: reward
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Tasks]
      summary: Update task
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - in: query
          name: title
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: description
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: reward
          required: true
          schema:
            type: integer
            minimum: 1
        - in: query
          name: status
          required: true
          schema:
            $ref: '#/components/schemas/TaskStatus'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /tasks/{id}/publish:
    post:
      tags: [Tasks]
      summary: Publish task
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Published task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /tasks/{id}/cancel:
    post:
      tags: [Tasks]
      summary: Cancel task
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Cancelled
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /bids:
    post:
      tags: [Bids]
      summary: Create bid
      parameters:
        - in: query
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Created bid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /bids/task/{taskId}:
    get:
      tags: [Bids]
      summary: List bids for task with pagination
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Bids list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bid'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /bids/{bidId}/select:
    post:
      tags: [Bids]
      summary: Select bid for task
      parameters:
        - in: path
          name: bidId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Selected
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /bids/task/{taskId}/complete:
    post:
      tags: [Bids]
      summary: Mark task as completed by executor
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Completed
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /payments/confirm:
    post:
      tags: [Payments]
      summary: Confirm task completion and perform payment
      parameters:
        - in: query
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment confirmed
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    Role:
      type: string
      enum: [CUSTOMER, EXECUTOR, ADMIN]

    TaskStatus:
      type: string
      enum: [CREATED, PUBLISHED, IN_PROGRESS, DONE, CONFIRMED, CANCELLED]

    BidStatus:
      type: string
      enum: [PENDING, ACCEPTED, REJECTED]

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        password:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        balance:
          type: integer
        blocked:
          type: boolean
      required: [id, email, password, role, balance, blocked]

    Task:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        reward:
          type: integer
        status:
          $ref: '#/components/schemas/TaskStatus'
        executorId:
          type: string
          format: uuid
      required: [title, description, reward, status]

    Bid:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/BidStatus'
      required: [id, taskId, status]

    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fromUser:
          type: integer
          format: int64
        toUser:
          type: integer
          format: int64
        amount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, fromUser, toUser, amount, createdAt]

    FieldViolation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required: [field, message]

    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
        requestId:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
      required: [timestamp, status, error, code, message, path, requestId]

    AuthUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum: [ CUSTOMER, EXECUTOR, ADMIN ]
        balance:
          type: integer
          format: int32
        blocked:
          type: boolean

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/AuthUserResponse'

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'